backend/requirements.txt
# Django 프로젝트 의존성 패키지

# Django 코어
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# 인증 및 보안
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
argon2-cffi==23.1.0
cryptography==41.0.7

# 데이터베이스
psycopg2-binary==2.9.8
redis==5.0.1
django-redis==5.4.0

# 태스크 큐
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# 파일 처리
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# API 문서화
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# 개발 도구
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# 테스팅
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# 보안 스캐닝
bandit==1.7.5
safety==2.3.5

# 환경 설정
python-decouple==3.8
python-dotenv==1.0.0

# 로깅 및 모니터링
sentry-sdk==1.38.0
django-health-check==3.17.0

# 데이터 검증
marshmallow==3.20.1
django-model-utils==4.3.1

# 성능 최적화
django-cache-memoize==0.1.10
django-cachalot==2.6.1

# 마이그레이션 도구
pyodbc==5.0.1  # SQL Server 연결용